+++ trie-db/CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 
+## [0.22.3] - 2021-01-28
+### Added
+- Decode with an iterator. [#121](https://github.com/paritytech/trie/pull/121)
+
+## [0.22.2] - 2021-01-05
+- Update `hashbrown` to 0.9. [#118](https://github.com/paritytech/trie/pull/118)
+
+## [0.22.1] - 2020-07-24
+- Use `ahash` feature of hashbrown. [#103](https://github.com/paritytech/trie/pull/103)
+
+## [0.22.0] - 2020-07-06
+- Update hashbrown to 0.8. [#97](https://github.com/paritytech/trie/pull/97)
+
+## [0.21.0] - 2020-06-04
+- Added `ALLOW_EMPTY` to `TrieLayout`. [#92](https://github.com/paritytech/trie/pull/92)
+
 ## [0.20.0] - 2020-02-07
 - Prefix iterator. [#39](https://github.com/paritytech/trie/pull/39)
 - Update trie-root to v0.16.0 and memory-db to v0.19.0 and reference-trie to v0.20.0 [#78](https://github.com/paritytech/trie/pull/78)